requirements.txt
requests
pynput
openai-whisper
sounddevice
fastapi[all]
numpy